item spwan modify
하나 없어지면 바로 생성되도록 수정
