+++ Ai_Project_Team_4/Assets/_Scripts/System/LerningSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 99810a467fb04404393789f328d08a65
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
