+++ Ai_Project_Team_4/Assets/_Scripts/System/Password.prefab
@@ -198,7 +198,7 @@ Transform:
   m_GameObject: {fileID: 7788402079656402207}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalPosition: {x: 0, y: 0, z: -1}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
